Large restructure of package
* Remove CWrapper submodule
* Remove export of ManagedHiGHS
* Export all C API
* Revise README
