+++ src/HiGHS.jl
@@ -1,24 +1,20 @@
 module HiGHS
 
-export ManagedHiGHS
-
 import HiGHS_jll: libhighs
 
-"""
-Wrapper for the HiGHS C library.
-"""
-module CWrapper
-
-import HiGHS.libhighs
-include(joinpath("wrapper", "libhighs_api.jl"))
-include(joinpath("wrapper", "libhighs_common.jl"))
-include(joinpath("wrapper", "api_helpers.jl"))
-
-end # module CWrapper
-
-include("c_model.jl")
-include("options.jl")
+include("wrapper/ctypes.jl")
+include("wrapper/libhighs_common.jl")
+include("wrapper/libhighs_api.jl")
 
 include("MOI_wrapper.jl")
+
+# HiGHS exports all `Highs_xxx` symbols. If you don't want all of these symbols
+# in your environment, then use `import HiGHS` instead of `using HiGHS`.
+
+for sym in names(@__MODULE__, all = true)
+    if startswith(string(sym), "Highs_")
+        @eval export $sym
+    end
+end
 
 end # module HiGHS
