+++ src/MOI_wrapper.jl
@@ -1,3 +1,57 @@
+
+mutable struct ManagedHiGHS
+    inner::Ptr{Cvoid}
+
+    function ManagedHiGHS()
+        mghs = new(
+            Highs_create()
+        )
+        # register the memory cleanup function
+        finalizer(mghs) do m
+            success = free_highs(m)
+            if !success
+                @warn "Memory free failure, possible leak."
+            end
+        end
+    end
+end
+
+"""
+Release references and free memory and return a boolean
+indicating the success of the operation.
+"""
+function free_highs(mhgs::ManagedHiGHS)
+    # Avoid double-free (ManagedHiGHS will set the pointers to NULL).
+    if mhgs.inner == C_NULL
+        return false
+    end
+    # only mhgs.inner is GC-protected during ccall!
+    GC.@preserve mhgs begin
+        Highs_destroy(mhgs.inner)
+    end
+    mhgs.inner = C_NULL
+    return true
+end
+
+"""
+    reset_model!(mhgs::ManagedHiGHS)
+
+Deletes the inner HiGHS model and recreates one.
+"""
+function reset_model!(mhgs::ManagedHiGHS)
+    # Avoid double-free (ManagedHiGHS will set the pointers to NULL).
+    if mhgs.inner == C_NULL
+        return false
+    end
+    # only mhgs.inner is GC-protected during ccall!
+    GC.@preserve mhgs begin
+        Highs_destroy(mhgs.inner)
+    end
+    mhgs.inner = Highs_create()
+    return true
+end
+
+
 import MathOptInterface
 const MOI = MathOptInterface
 
@@ -18,8 +72,8 @@ function MOI.empty!(o::Optimizer)
 end
 
 function MOI.is_empty(o::Optimizer)
-    return CWrapper.Highs_getNumRows(o.model.inner) == 0 &&
-      CWrapper.Highs_getNumCols(o.model.inner) == 0 &&
+    return Highs_getNumRows(o.model.inner) == 0 &&
+      Highs_getNumCols(o.model.inner) == 0 &&
       o.objective_sense == MOI.FEASIBILITY_SENSE &&
       o.objective_constant == 0.0
 end
@@ -31,18 +85,18 @@ function MOI.optimize!(o::Optimizer)
             error("Feasibility sense with non-constant objective function, set the sense to min/max, the objective to 0 or reset the sense to feasibility to erase the objective")
         end
     end
-    CWrapper.Highs_run(o.model.inner)
+    Highs_run(o.model.inner)
     return
 end
 
 function MOI.add_variable(o::Optimizer)
-    _ = CWrapper.Highs_addCol(o.model.inner, 0.0, -Inf, Inf, Cint(0), Cint[], Cint[])
+    _ = Highs_addCol(o.model.inner, 0.0, -Inf, Inf, Cint(0), Cint[], Cint[])
     col_idx = MOI.get(o, MOI.NumberOfVariables()) - 1
     return MOI.VariableIndex(col_idx)
 end
 
 function MOI.add_constrained_variable(o::Optimizer, set::S) where {S <: MOI.Interval}
-    _ = CWrapper.Highs_addCol(o.model.inner, 0.0, Cdouble(set.lower), Cdouble(set.upper), Cint(0), Cint[], Cint[])
+    _ = Highs_addCol(o.model.inner, 0.0, Cdouble(set.lower), Cdouble(set.upper), Cint(0), Cint[], Cint[])
     col_idx = MOI.get(o, MOI.NumberOfVariables()) - 1
     return (MOI.VariableIndex(col_idx), MOI.ConstraintIndex{MOI.SingleVariable, MOI.Interval{Float64}}(col_idx))
 end
@@ -51,12 +105,12 @@ MOI.supports_constraint(::Optimizer, ::MOI.SingleVariable, ::MOI.Interval) = tru
 
 function MOI.add_constraint(o::Optimizer, sg::MOI.SingleVariable, set::MOI.Interval)
     var_idx = Cint(sg.variable.value)
-    _ = CWrapper.Highs_changeColBounds(o.model.inner, var_idx, Cdouble(set.lower), Cdouble(set.upper))
+    _ = Highs_changeColBounds(o.model.inner, var_idx, Cdouble(set.lower), Cdouble(set.upper))
     return
 end
 
 function MOI.get(o::Optimizer, ::MOI.NumberOfConstraints{MOI.ScalarAffineFunction{Float64}, MOI.Interval{Float64}})
-    nrows = CWrapper.Highs_getNumRows(o.model.inner)
+    nrows = Highs_getNumRows(o.model.inner)
     return Int(nrows)
 end
 
@@ -75,7 +129,7 @@ function MOI.add_constraint(o::Optimizer, func::MOI.ScalarAffineFunction, set::M
     col_indices_ptr = pointer(col_indices)
     lower = convert(Cdouble, set.lower - func.constant)
     upper = convert(Cdouble, set.upper - func.constant)
-    row_idx = CWrapper.Highs_addRow(o.model.inner, lower, upper, Cint(number_nonzeros), col_indices_ptr, coefficients_ptr)
+    row_idx = Highs_addRow(o.model.inner, lower, upper, Cint(number_nonzeros), col_indices_ptr, coefficients_ptr)
     return MOI.ConstraintIndex{typeof(func), typeof(set)}(row_idx)
 end
 
@@ -86,17 +140,17 @@ MOI.supports(::Optimizer, param::MOI.RawParameter) = true
 
 # setting HiGHS options
 function MOI.set(o::Optimizer, param::MOI.RawParameter, value)
-    CWrapper.set_option(o.model.inner, param.name, value)
+    set_option(o.model.inner, param.name, value)
     return nothing
 end
 
 function MOI.get(o::Optimizer, param::MOI.RawParameter)
     return if param.name in options_string
-        CWrapper.get_option(o.model.inner, param.name, String)
+        get_option(o.model.inner, param.name, String)
     elseif param.name in options_int
-        CWrapper.get_option(o.model.inner, param.name, Int)
+        get_option(o.model.inner, param.name, Int)
     elseif param.name in options_double
-        CWrapper.get_option(o.model.inner, param.name, Float64)        
+        get_option(o.model.inner, param.name, Float64)
     else
         throw(ArgumentError("Parameter $(param.name) is not supported"))
     end
@@ -131,7 +185,7 @@ MOI.supports(::Optimizer, ::MOI.VariableName, ::Type{MOI.VariableIndex}) = true
 MOI.get(o::Optimizer, ::MOI.RawSolver) = o.model
 
 function MOI.get(o::Optimizer, ::MOI.ResultCount)
-    status = CWrapper.Highs_getModelStatus(o.model.inner)
+    status = Highs_getModelStatus(o.model.inner)
     status == 9 ? 1 : 0
 end
 
@@ -141,7 +195,7 @@ end
 
 function MOI.set(o::Optimizer, ::MOI.ObjectiveSense, sense::MOI.OptimizationSense)
     sense_code = sense == MOI.MAX_SENSE ? Cint(-1) : Cint(1)
-    _ = CWrapper.Highs_changeObjectiveSense(o.model.inner, sense_code)
+    _ = Highs_changeObjectiveSense(o.model.inner, sense_code)
     o.objective_sense = sense
     # if feasibility sense set, erase the function
     if sense == MOI.FEASIBILITY_SENSE
@@ -151,7 +205,7 @@ function MOI.set(o::Optimizer, ::MOI.ObjectiveSense, sense::MOI.OptimizationSens
 end
 
 function MOI.get(o::Optimizer, ::MOI.NumberOfVariables)
-    return Int(CWrapper.Highs_getNumCols(o.model.inner))
+    return Int(Highs_getNumCols(o.model.inner))
 end
 
 function MOI.get(o::Optimizer, ::MOI.ListOfVariableIndices)
@@ -172,7 +226,7 @@ function MOI.set(o::Optimizer, ::MOI.ObjectiveFunction{F}, func::F) where {F <: 
     end
     coefficients_ptr = pointer(coefficients)
     mask = pointer(ones(Cint, total_ncols))
-    CWrapper.Highs_changeColsCostByMask(o.model.inner, mask, coefficients_ptr)
+    Highs_changeColsCostByMask(o.model.inner, mask, coefficients_ptr)
     o.objective_constant = MOI.constant(func)
     return
 end
@@ -181,7 +235,7 @@ function MOI.get(o::Optimizer, ::MOI.ObjectiveFunction{MOI.ScalarAffineFunction{
     ncols = MOI.get(o, MOI.NumberOfVariables())
     num_cols = Ref{Cint}(0)
     costs = Vector{Cdouble}(undef, ncols)
-    _ = CWrapper.Highs_getColsByRange(
+    _ = Highs_getColsByRange(
         o.model.inner,
         Cint(0), Cint(ncols-1), # column range
         num_cols, costs,
@@ -202,13 +256,13 @@ end
 function MOI.get(o::Optimizer, attr::MOI.ObjectiveValue)
     MOI.check_result_index_bounds(o, attr)
     value = Ref{Cdouble}()
-    CWrapper.Highs_getHighsDoubleInfoValue(o.model.inner, "objective_function_value", value);
+    Highs_getHighsDoubleInfoValue(o.model.inner, "objective_function_value", value);
     return o.objective_constant + value[]
 end
 
 function MOI.get(o::Optimizer, ::MOI.SimplexIterations)
     simplex_iteration_count = Ref{Cint}(0)
-    CWrapper.Highs_getHighsIntInfoValue(o.model.inner, "simplex_iteration_count", simplex_iteration_count)
+    Highs_getHighsIntInfoValue(o.model.inner, "simplex_iteration_count", simplex_iteration_count)
     return Int(simplex_iteration_count[])
 end
 
@@ -236,11 +290,83 @@ end
 
 function MOI.get(o::Optimizer, ::MOI.TimeLimitSec)
     value = Ref{Cdouble}()
-    CWrapper.Highs_getHighsDoubleOptionValue(o.model.inner, "time_limit", value)
+    Highs_getHighsDoubleOptionValue(o.model.inner, "time_limit", value)
     return value[]
 end
 
 function MOI.set(o::Optimizer, ::MOI.TimeLimitSec, value)
-    CWrapper.set_option(o.model.inner, "time_limit", Float64(value))
+    set_option(o.model.inner, "time_limit", Float64(value))
     return
+end
+
+# Completes the generated wrapper with default methods
+
+# TODO update these to leverage dispatch
+
+"""
+    set_option(highs, option, value::T)
+
+Sets the option to the value, dispatches on
+the value type T to call appropriate HiGHS function.
+"""
+function set_option end
+
+function set_option(highs, option, value::Integer)
+    Highs_setHighsIntOptionValue(highs, option, Cint(value))
+end
+
+function set_option(highs, option, value::Bool)
+    Highs_setHighsBoolOptionValue(highs, option, Cint(value))
+end
+
+function set_option(highs, option, value::AbstractFloat)
+    Highs_setHighsDoubleOptionValue(highs, option, Cdouble(value))
+end
+
+function set_option(highs, option, value::String)
+    Highs_setHighsStringOptionValue(highs, option, value)
+end
+
+# add all options methods here
+
+const options_string = ["presolve", "solver", "parallel"]
+const options_int = ["simplex_strategy", "simplex_iteration_limit", "highs_min_threads", "message_level"]
+const options_double = ["time_limit"]
+
+"""
+    get_option(highs, option, T)
+
+Gets the option of type T on the HiGHS side.
+"""
+function get_option end
+
+function get_option(highs, option, ::Type{T}) where {T <: Integer}
+    value = Ref{Cint}(0)
+    Highs_getHighsIntOptionValue(highs, option, value)
+    return T(value[])
+end
+
+function get_option(highs, option, ::Type{Bool})
+    value = Ref{Cint}(0)
+    Highs_getHighsBoolOptionValue(highs, option, value)
+    return Bool(value[])
+end
+
+function get_option(highs, option, ::Type{T}) where {T <: AbstractFloat}
+    value = Ref{Cdouble}()
+    Highs_getHighsDoubleOptionValue(highs, option, value)
+    return T(value[])
+end
+
+function get_option(highs, option, ::Type{String})
+    v = Vector{Cchar}(undef, 100)
+    p = pointer(v)
+    Highs_getHighsStringOptionValue(highs, option, p)
+    GC.@preserve v s = unsafe_string(p)
+    return s
+end
+
+# convenience to accomodate/ignore scale argument
+function Highs_getModelStatus(highs)
+    Highs_getModelStatus(highs, Cint(0))
 end
