--- src/options.jl
@@ -1,5 +1,0 @@
-# add all options methods here
-
-const options_string = ["presolve", "solver", "parallel"]
-const options_int = ["simplex_strategy", "simplex_iteration_limit", "highs_min_threads", "message_level"]
-const options_double = ["time_limit"]
